+++ pkgs/cli/CHANGELOG.md
@@ -1,5 +1,19 @@
 # pgflow
 
+## 0.3.1
+
+### Patch Changes
+
+- Updated dependencies [d08fd2d]
+  - @pgflow/core@0.3.1
+
+## 0.3.0
+
+### Patch Changes
+
+- Updated dependencies [c3653fa]
+  - @pgflow/core@0.3.0
+
 ## 0.2.6
 
 ### Patch Changes